Move common typescript types to a proper location
