+++ specifyweb/frontend/js_src/lib/cache.ts
@@ -1,6 +1,6 @@
 import type { CacheDefinitions } from './cachedefinitions';
 import { safeToTrim } from './cachedefinitions';
-import type { R, RA } from './components/wbplanview';
+import type { R, RA } from './types';
 
 // Determines how persistent bucket's storage would be
 type BucketType =
