+++ specifyweb/frontend/js_src/lib/components/common.tsx
@@ -7,8 +7,11 @@ import { spanNumber } from '../wbplanviewhelper';
 import dataModelStorage from '../wbplanviewmodel';
 
 const MAX_HUE = 360;
+
 const getHue = spanNumber(
+  // eslint-disable-next-line unicorn/prefer-code-point
   'a'.charCodeAt(0) * 2,
+  // eslint-disable-next-line unicorn/prefer-code-point
   'z'.charCodeAt(0) * 2,
   0,
   MAX_HUE
@@ -30,6 +33,7 @@ export function TableIcon({
       />
     );
 
+  // eslint-disable-next-line unicorn/prefer-code-point
   const colorHue = getHue(tableName.charCodeAt(0) + tableName.charCodeAt(0));
   const color = `hsl(${colorHue}, 70%, 50%)`;
   return (
@@ -43,12 +47,12 @@ export function TableIcon({
   );
 }
 
-export const TableIconUndefined = (
+export const tableIconUndefined = (
   <span className="table-icon table-icon-undefined">⃠</span>
 );
 
-export const TableIconSelected = (
+export const tableIconSelected = (
   <span className="table-icon table-icon-selected">✓</span>
 );
 
-export const TableIconEmpty = <span className="table-icon table-icon-emtpy" />;
+export const tableIconEmpty = <span className="table-icon table-icon-emtpy" />;
