+++ specifyweb/frontend/js_src/lib/leafletconfig.ts
@@ -3,7 +3,7 @@
  *
  */
 
-import type { IR, RA, RR } from './components/wbplanview';
+import type { IR, RA, RR } from './types';
 import type { MappingPath } from './components/wbplanviewmapper';
 import L from './leafletextend';
