+++ specifyweb/frontend/js_src/lib/legacytypes.ts
@@ -1,4 +1,4 @@
-import type { IR, RA } from './components/wbplanview';
+import type { IR, RA } from './types';
 import type { RelationshipType } from './components/wbplanviewmapper';
 
 export interface SchemaModelTableField {
