+++ specifyweb/frontend/js_src/lib/lifemapperhelper.ts
@@ -1,4 +1,4 @@
-import type { R, RA } from './components/wbplanview';
+import type { R, RA } from './types';
 import type { MappingPath } from './components/wbplanviewmapper';
 import type { Field, LocalityData } from './leafletutils';
 import { findArrayDivergencePoint } from './wbplanviewhelper';
