+++ specifyweb/frontend/js_src/lib/lifemapperreducer.ts
@@ -2,7 +2,7 @@ import type { Action } from 'typesafe-reducer';
 import { generateReducer } from 'typesafe-reducer';
 
 import type { States } from './components/lifemapperstate';
-import type { IR, RA, RR } from './components/wbplanview';
+import type { IR, RA, RR } from './types';
 import type { LayerConfig, MarkerGroups } from './leaflet';
 import type { MessageTypes } from './lifemapperconfig';
