+++ specifyweb/frontend/js_src/lib/uploadplantomappingstree.ts
@@ -1,4 +1,4 @@
-import type { IR, R, RA } from './components/wbplanview';
+import type { IR, R, RA } from './types';
 import type { MappingPath } from './components/wbplanviewmapper';
 import { defaultColumnOptions } from './wbplanviewlinesgetter';
 import { formatReferenceItem, formatTreeRank } from './wbplanviewmappinghelper';
