+++ specifyweb/frontend/js_src/lib/wbplanviewheaderhelper.ts
@@ -1,4 +1,4 @@
-import type { RA } from './components/wbplanview';
+import type { RA } from './types';
 import type { MappingLine } from './components/wbplanviewmapper';
 import wbText from './localization/workbench';
 import { generateMappingPathPreview } from './wbplanviewmappingpreview';
