+++ specifyweb/frontend/js_src/lib/wbplanviewmodelconfig.ts
@@ -1,4 +1,4 @@
-import type { IR } from './components/wbplanview';
+import type { IR } from './types';
 import type { RelationshipType } from './components/wbplanviewmapper';
 import schema from './schema';
