+++ specifyweb/frontend/js_src/lib/wbplanviewmodelhelper.ts
@@ -6,7 +6,7 @@
  *
  */
 
-import type { IR, RA } from './components/wbplanview';
+import type { IR, RA } from './types';
 import type {
   MappingPath,
   RelationshipType,
