+++ specifyweb/frontend/js_src/lib/wbworker.d.ts
@@ -1,0 +1,7 @@
+declare module 'worker-loader!*' {
+  class WebpackWorker extends Worker {
+    constructor();
+  }
+
+  export default WebpackWorker;
+}
